Completion list item selector method - ability to set for each editor instances
