+++ src/vs/editor/common/languages.ts
@@ -693,11 +693,6 @@ export interface CompletionList {
 	duration?: number;
 }
 
-export type CompletionItemInfo = { completion: CompletionItem; word?: string };
-
-export type CompletionListItemSelectionMethod
-	= (isAuto: boolean, selectionIndex: number, items: CompletionItemInfo[]) => number;
-
 /**
  * How a suggest provider was triggered.
  */
