+++ src/vs/editor/standalone/browser/standaloneLanguages.ts
@@ -22,7 +22,6 @@ import { IMonarchLanguage } from 'vs/editor/standalone/common/monarch/monarchTyp
 import { IStandaloneThemeService } from 'vs/editor/standalone/common/standaloneTheme';
 import { IMarkerData, IMarkerService } from 'vs/platform/markers/common/markers';
 import { ILanguageFeaturesService } from 'vs/editor/common/services/languageFeatures';
-import { IEditorCompletionService } from 'vs/editor/common/services/editorCompletionService';
 import { LanguageSelector } from 'vs/editor/common/languageSelector';
 
 /**
@@ -657,13 +656,6 @@ export function registerInlayHintsProvider(languageSelector: LanguageSelector, p
 }
 
 /**
- * Register a custom completion list item selection method.
- */
-export function registerCompletionListItemSelectorMethod(method: languages.CompletionListItemSelectionMethod): void {
-	const editorCompletionService = StandaloneServices.get(IEditorCompletionService);
-	editorCompletionService.registerCompletionListItemSelectorMethod(method);
-}
-/**
  * Contains additional diagnostic information about the context in which
  * a [code action](#CodeActionProvider.provideCodeActions) is run.
  */
@@ -754,7 +746,6 @@ export function createMonacoLanguagesAPI(): typeof monaco.languages {
 		registerDocumentRangeSemanticTokensProvider: <any>registerDocumentRangeSemanticTokensProvider,
 		registerInlineCompletionsProvider: <any>registerInlineCompletionsProvider,
 		registerInlayHintsProvider: <any>registerInlayHintsProvider,
-		registerCompletionListItemSelectorMethod: <any>registerCompletionListItemSelectorMethod,
 
 		// enums
 		DocumentHighlightKind: standaloneEnums.DocumentHighlightKind,
